api/requirements.txt
flask
flask_cors
python-dotenv
langchain
openai
faiss-cpu
pdfplumber
langchain_community
werkzeug
langchain_openai